Create fragment for weather list. Add navigation support.
