+++ app/src/main/res/navigation/nav_weather.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    app:startDestination="@id/weather_fragment">
+
+    <fragment
+        android:id="@+id/weather_fragment"
+        android:name="org.eldorado.simpleweather.ui.WeatherListFragment"
+        tools:layout="@layout/fragment_weather">
+    </fragment>
+
+</navigation>
